Refac: ControlBox, Card, StudyHeader 컴포넌트 분리
