+++ src/components/study/index.ts
@@ -1,3 +1,6 @@
 export { default as ToggleGuideCard } from './ToggleGuideCard';
 export { default as GhostCard } from './GhostCard';
 export { default as EmptyCard } from './EmptyCard';
+export { default as StudyHeader } from './StudyHeader';
+export { default as Card } from './Card';
+export { default as ControlBox } from './ControlBox';
